Second attempt at refactor
