+++ src/gb/auth/tokenHandler.ts
@@ -1,4 +1,4 @@
-import { GameBusClient } from '../gameBusClient';
+import { GameBusClient } from '../gbClient';
 
 /**
  * Token handler that can retrieve and refresh tokens if needed
